+++ NanoXLSX/LowLevel/XlsxWriter.cs
@@ -1216,7 +1216,7 @@ namespace NanoXLSX.LowLevel
                     {
                         throw new FormatException("The number format style component with the ID " + Utils.ToString(item.CustomFormatID) + " cannot be null or empty");
                     }
-                    String customFormat = NumberFormat.EscapeFormatCode(item.CustomFormatCode);
+                    string customFormat = NumberFormat.EscapeFormatCode(item.CustomFormatCode);
                     sb.Append("<numFmt formatCode=\"").Append(EscapeXmlAttributeChars(customFormat)).Append("\" numFmtId=\"").Append(item.CustomFormatID.ToString("G", culture)).Append("\"/>");
                 }
             }